solved old gem problem and added Dockerfile
the gem em-twitter is not able to be built on ruby 2.2, I open an PR here: https://github.com/tweetstream/em-twitter/pull/22. But it seems that eventmachine 1.0.4 wont work on ruby 2.1. I use my repo as gem as a work-around ... also I added files for docker in order to deploy automatically